refactored Nick's work and integrated it into the existing workflow
