+++ src/main/java/edu/uco/cs/v2c/desktop/linux/ui/EditCommand.java
@@ -1,4 +1,4 @@
-package edu.uco.cs.v2c.desktop.linux.nick.view;
+package edu.uco.cs.v2c.desktop.linux.ui;
 
 import java.awt.BorderLayout;
 import java.awt.Container;
@@ -13,7 +13,7 @@ import javax.swing.JPanel;
 import javax.swing.JTextField;
 import javax.swing.border.Border;
 
-import edu.uco.cs.v2c.desktop.linux.nick.controller.EditEventListener;
+import edu.uco.cs.v2c.desktop.linux.command.EditEventListener;
 
 public class EditCommand {
